refactor #260: expunge proto::ClientResult from user-facing API
Replace proto::ClientResult with core::KvEntry as the return type of
GrpcClient::get_with_policy() and get_multi_with_policy().  Previously
these power-user methods leaked a protobuf-generated type into the public
API surface despite the rest of the read API (ClientApi::get,
get_linearizable, etc.) already returning plain Bytes.

Changes:
- client_ext.rs: into_read_results() now returns Vec<Option<KvEntry>>
- grpc_client.rs: get_with_policy / get_multi_with_policy return KvEntry
- lib.rs: protocol module replaces ClientResult export with KvEntry
- examples/client-usage-standalone: updated to KvEntry pattern match

Tests (TDD -- red before green):
- test_into_read_results_success: explicit Vec<Option<KvEntry>> annotation
- test_into_read_results_multiple_entries_are_kv_entries (new)
- test_get_with_policy_returns_native_kv_entry (new)
- test_get_multi_with_policy_returns_native_kv_entries (new)

proto::ClientResult survives only inside d-engine-proto's wire layer.
All user-visible read results are now proto-free native Rust types.

+++ d-engine-core/src/raft_role/leader_state_test/fatal_error_test.rs
@@ -10,12 +10,12 @@ use crate::MockTypeConfig;
 use crate::RaftEvent;
 use crate::ReadConsistencyPolicy;
 use crate::RoleEvent;
+use crate::client::ClientReadRequest;
 use crate::maybe_clone_oneshot::MaybeCloneOneshot;
 use crate::maybe_clone_oneshot::RaftOneshot;
 use crate::role_state::RaftRoleState;
 use crate::test_utils::MockBuilder;
 use bytes::Bytes;
-use d_engine_proto::client::ClientReadRequest;
 use std::sync::Arc;
 use tokio::sync::{mpsc, watch};
 use tonic::Code;
@@ -145,7 +145,7 @@ async fn test_fatal_error_drains_all_pending_queues() {
     let lin_req = ClientReadRequest {
         client_id: 1,
         keys: vec![Bytes::from_static(b"k1")],
-        consistency_policy: Some(ReadConsistencyPolicy::LinearizableRead as i32),
+        consistency_policy: Some(ReadConsistencyPolicy::LinearizableRead),
     };
     leader.push_client_cmd(ClientCmd::Read(lin_req, lin_tx), &ctx);
     assert_eq!(
@@ -159,7 +159,7 @@ async fn test_fatal_error_drains_all_pending_queues() {
     let lease_req = ClientReadRequest {
         client_id: 2,
         keys: vec![Bytes::from_static(b"k2")],
-        consistency_policy: Some(ReadConsistencyPolicy::LeaseRead as i32),
+        consistency_policy: Some(ReadConsistencyPolicy::LeaseRead),
     };
     leader.test_update_lease_timestamp();
     leader.push_client_cmd(ClientCmd::Read(lease_req, lease_tx), &ctx);
@@ -174,7 +174,7 @@ async fn test_fatal_error_drains_all_pending_queues() {
     let ev_req = ClientReadRequest {
         client_id: 3,
         keys: vec![Bytes::from_static(b"k3")],
-        consistency_policy: Some(ReadConsistencyPolicy::EventualConsistency as i32),
+        consistency_policy: Some(ReadConsistencyPolicy::EventualConsistency),
     };
     leader.push_client_cmd(ClientCmd::Read(ev_req, ev_tx), &ctx);
     assert_eq!(
